requirements.txt
Flask
requests
python-dotenv
Flask-SQLAlchemy
werkzeug
pymysql
Flask-WTF==1.2.1